+++ zrtp/libzrtpcpp/ZRtp.h
@@ -1422,8 +1422,8 @@ private:
       */
      void setClientId(std::string id, HelloPacketVersion* hpv);
 
-     void log(uint severity, const char *obj, const char *fmt, ...);
-     void vlog(uint severity, const char *obj, const char *fmt, va_list argp);
+     void log(uint8_t severity, const char *obj, const char *fmt, ...);
+     void vlog(uint8_t severity, const char *obj, const char *fmt, va_list argp);
 };
 
 /**
